requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
openai>=1.30,<2.0
pdfplumber>=0.11,<1.0
pytesseract>=0.3.10,<1.0
Pillow>=10.3,<12.0
python-docx>=1.1,<2.0
rapidfuzz>=3.9,<4.0
fpdf2>=2.7,<3.0
pyyaml>=6.0,<7.0
pydantic>=2.7,<3.0
python-dotenv>=1.0,<2.0
pytest>=8.2,<9.0
httpx>=0.27,<1.0
